+++ src/codegen/stmnt.rs
@@ -2,7 +2,7 @@ use cranelift::prelude::{FunctionBuilder, InstBuilder};
 use cranelift_module::Module;
 
 use crate::typechecker::{
-    stmnt::{Block, Stmnt},
+    stmnt::{Block, Return, Stmnt},
     value::FuncCall,
 };
 
@@ -22,12 +22,33 @@ pub fn compile_func_call(codegen: &mut Codegen, builder: &mut FunctionBuilder, c
     builder.ins().call(func, &args);
 }
 
-pub fn compile_statement(codegen: &mut Codegen, builder: &mut FunctionBuilder, stmnt: &Stmnt) {
+pub fn compile_return(codegen: &mut Codegen, builder: &mut FunctionBuilder, ret: &Return) {
+    let mut args = Vec::new();
+
+    ret.value.iter().for_each(|value| {
+        args.push(super::value::compile_value(codegen, builder, value));
+    });
+
+    builder.ins().return_(&args);
+}
+
+/// Returns true if the statement returns.
+pub fn compile_statement(
+    codegen: &mut Codegen,
+    builder: &mut FunctionBuilder,
+    stmnt: &Stmnt,
+) -> bool {
     match stmnt {
         Stmnt::FuncCall(func_call) => {
             compile_func_call(codegen, builder, func_call);
         }
+        Stmnt::Return(ret) => {
+            compile_return(codegen, builder, ret);
+            return true;
+        }
     }
+
+    false
 }
 
 pub fn compile_block(codegen: &mut Codegen, builder: &mut FunctionBuilder, block: &Block) {
@@ -37,5 +58,5 @@ pub fn compile_block(codegen: &mut Codegen, builder: &mut FunctionBuilder, block
 
     // TODO: check if returned
 
-    builder.ins().return_(&[]);
+    // builder.ins().return_(&[]);
 }