+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Language
 - [feat (#3): [syntax/parser] implement string escapes](https://github.com/amp-lang/amp/issues/3)
 - [feat (#5): [syntax/parser, typechecker, codegen] decimal integer literals](https://github.com/amp-lang/amp/issues/5)
+- [feat (#4): [syntax/parser, typechecker, codegen] return statements](https://github.com/amp-lang/amp/issues/4)
 
 ### Compiler
 - feat: Basic stopwatch benchmark for Amp compile times.